requirements.txt
mcp>=0.9.0
stravalib>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0